requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.2
Flask-Migrate==4.0.5
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
frozendict==2.3.8
frozenlist==1.4.0
greenlet==3.0.1
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
multitasking==0.0.11
mysqlclient==2.2.0
numpy==1.26.1
openai==0.28.1
optional-django==0.3.0
packaging==23.2
pandas==2.1.2
peewee==3.17.0
plotly==5.18.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
react==4.3.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
Werkzeug==2.3.7
yarl==1.9.2
yfinance==0.2.31
